+++ tests/integration/test_bytes.rs
@@ -13,8 +13,9 @@ fn read_bytes_from_path(path: &str) -> eyre::Result<Vec<ParsedEvent>> {
     let mut reader =
         new_filesystem_reader(path, ConnectorMode::Static, ReadMethod::Full, "*", false)?;
     let mut parser = IdentityParser::new(
-        vec!["data".to_string()],
+        &["data".to_string()],
         false,
+        None,
         KeyGenerationPolicy::PreferMessageKey,
         SessionType::Native,
     );